requirements.txt
langchain
langchain-community
langchain-groq
langchainhub
streamlit
wikipedia
arxiv
sentence-transformers
chromadb
faiss-cpu
python-dotenv